Update manifest.json
Simply I removed one character - the trailing possessive apostrophe (') from the end of the description attribute. It does not belong or was simply a typo. I was tempted to also change "Analyses" to "Analyzes" since that's the form I'm used to, but did not since both are legit (http://grammarist.com/spelling/analyse-analyze/).